split dashboard user loading for web and desktop
